# .factory/api-contracts.yaml
openapi: 3.0.3
info:
  title: Dom'Atic API
  description: |
    Local IoT Hub — centralized device management, real-time monitoring,
    automation rules, and security alerting. All endpoints (except auth)
    require a valid Bearer JWT token obtained via POST /auth/login.
  version: 1.0.0
  contact:
    name: Dom'Atic
    email: admin@domatic.local
servers:
  - url: http://domatic.local:8000
    description: Local LAN server
  - url: http://localhost:8000
    description: Development

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access JWT (15 min). Obtain via POST /auth/login.
        Refresh via POST /auth/refresh using the HttpOnly cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    PaginatedMeta:
      type: object
      properties:
        total:    { type: integer, example: 142 }
        page:     { type: integer, example: 1 }
        per_page: { type: integer, example: 50 }
        pages:    { type: integer, example: 3 }

    # --- Error ---
...
    ProtocolType:
      type: string
      enum: [wifi, zigbee, zwave, ble, matter, ethernet, unknown]
    ChipType:
      type: string
      enum: [esp8266, esp32, esp32s2, esp32s3, esp32c3, bk7231, bk7231n, rtl8710, unknown]
    FirmwareType:
      type: string
      enum: [tasmota, esphome, openbk, openshock, original, custom, unknown]
    SecuritySeverity:
      type: string
      enum: [info, low, medium, high, critical]
    FirmwareUpdateStatus:
      type: string
      enum: [pending, in_progress, success, failed, rolled_back]

    # --- Auth ---
...
    Room:
      type: object
      properties:
        id:         { type: string, format: uuid }
        name:       { type: string, example: Living Room }
        icon:       { type: string, example: sofa, nullable: true }
        floor:      { type: integer, example: 0 }
        vlan_id:    { type: integer, nullable: true }
        color:      { type: string, example: '#3B82F6', nullable: true }
        device_count: { type: integer, example: 4 }
        online_count: { type: integer, example: 3 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RoomCreate:
      type: object
      required: [name]
      properties:
        name:    { type: string, maxLength: 100, example: Kitchen }
        icon:    { type: string, maxLength: 50, nullable: true }
        floor:   { type: integer, default: 0 }
        vlan_id: { type: integer, nullable: true }
        color:   { type: string, pattern: '^#[0-9A-Fa-f]{6}$', nullable: true }
    RoomUpdate:
      type: object
      properties:
        name:    { type: string, maxLength: 100 }
        icon:    { type: string, maxLength: 50, nullable: true }
        floor:   { type: integer }
        vlan_id: { type: integer, nullable: true }
        color:   { type: string, pattern: '^#[0-9A-Fa-f]{6}$', nullable: true }

    # --- Devices ---
    Device:
      type: object
      properties:
        id:                        { type: string, format: uuid }
        name:                      { type: string }
        room_id:                   { type: string, format: uuid, nullable: true }
        room_name:                 { type: string, nullable: true }
        protocol:                  { $ref: '#/components/schemas/ProtocolType' }
        chip_type:                 { $ref: '#/components/schemas/ChipType' }
        ip_address:                { type: string, format: ipv4, nullable: true }
        mac_address:               { type: string, example: 'AA:BB:CC:DD:EE:FF', nullable: true }
        firmware_type:             { $ref: '#/components/schemas/FirmwareType' }
        firmware_version:          { type: string, nullable: true }
        firmware_latest:           { type: string, nullable: true }
        firmware_update_available: { type: boolean }
        mqtt_topic:                { type: string, nullable: true }
        mqtt_lwt_topic:            { type: string, nullable: true }
        is_online:                 { type: boolean }
        is_cloud_dependent:        { type: boolean }
        is_blocked:                { type: boolean }
        blocked_reason:            { type: string, nullable: true }
        cloud_attempts:            { type: integer }
        last_seen:                 { type: string, format: date-time, nullable: true }
        metadata:                  { type: object, additionalProperties: true }
        unresolved_security_events: { type: integer }
        created_at:                { type: string, format: date-time }
        updated_at:                { type: string, format: date-time }
...
    SecurityEvent:
      type: object
      properties:
        id:               { type: string, format: uuid }
        device_id:        { type: string, format: uuid, nullable: true }
        device_name:      { type: string, nullable: true }
        event_type:       { type: string }
        description:      { type: string }
        severity:         { $ref: '#/components/schemas/SecuritySeverity' }
        source_ip:        { type: string, nullable: true }
        destination_ip:   { type: string, nullable: true }
        destination_host: { type: string, nullable: true }
        blocked:          { type: boolean }
        resolved:         { type: boolean }
        resolved_at:      { type: string, format: date-time, nullable: true }
        resolution_note:  { type: string, nullable: true }
        raw_data:         { type: object, additionalProperties: true }
        created_at:       { type: string, format: date-time }
...
    ResolveEventRequest:
      type: object
      required: [resolution_note]
      properties:
        resolution_note: { type: string, minLength: 1 }

    # --- Automations ---
...
    MqttMessage:
      type: object
      properties:
        id:           { type: integer }
        topic:        { type: string }
        payload:      { type: string, nullable: true }
        payload_json: { type: object, nullable: true }
        qos:          { type: integer, enum: [0, 1, 2] }
        retained:     { type: boolean }
        device_id:    { type: string, format: uuid, nullable: true }
        device_name:  { type: string, nullable: true }
        received_at:  { type: string, format: date-time }
    MqttPublishRequest:
      type: object
      required: [topic, payload]
      properties:
        topic:    { type: string, maxLength: 512 }
        payload:  { type: string }
        qos:      { type: integer, enum: [0, 1, 2], default: 0 }
        retain:   { type: boolean, default: false }

    # --- Firmware ---
...
    FirmwareUpdateLog:
      type: object
      required: [to_firmware, to_version]
      properties:
        from_firmware: { $ref: '#/components/schemas/FirmwareType' }
        to_firmware:   { $ref: '#/components/schemas/FirmwareType' }
        from_version:  { type: string, nullable: true }
        to_version:    { type: string }
        status:        { $ref: '#/components/schemas/FirmwareUpdateStatus' }
        notes:         { type: string, nullable: true }
        metadata:      { type: object, additionalProperties: true, default: {} }
    FirmwareUpdateRecord:
      allOf:
        - $ref: '#/components/schemas/FirmwareUpdateLog'
        - type: object
          properties:
            id:          { type: string, format: uuid }
            device_id:   { type: string, format: uuid }
            device_name: { type: string }
            created_at:  { type: string, format: date-time }
            updated_at:  { type: string, format: date-time }

    # --- Dashboard ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /rooms:
    get:
      tags: [Rooms]
      summary: List all rooms with device counts
      responses:
        '200':
          description: Room list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'

    post:
      tags: [Rooms]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreate'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '409':
          description: Room name already exists
  /rooms/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }

    get:
      tags: [Rooms]
      summary: Get room detail with devices
      responses:
        '200':
          description: Room with device list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Room'
                  - type: object
                    properties:
                      devices:
                        type: array
                        items: { $ref: '#/components/schemas/Device' }

    patch:
      tags: [Rooms]
      summary: Update room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdate'
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'

    delete:
      tags: [Rooms]
      summary: Delete room (devices become roomless, not deleted)
      responses:
        '204':
          description: Deleted

  # =========================================================================
  # SECURITY
  # =========================================================================
  /security/events:
    get:
      tags: [Security]
      summary: List security events
      parameters:
        - in: query
          name: severity
          schema: { $ref: '#/components/schemas/SecuritySeverity' }
        - in: query
          name: resolved
          schema: { type: boolean, default: false }
        - in: query
          name: device_id
          schema: { type: string, format: uuid }
        - in: query
          name: event_type
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Paginated security events
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SecurityEvent' }
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'
...
  /security/events/{id}/resolve:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Security]
      summary: Mark a security event as resolved
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveEventRequest'
      responses:
        '200':
          description: Event resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityEvent'
        '400':
          description: Event already resolved
        '404':
          description: Not found
...
  /mqtt/messages:
    get:
      tags: [MQTT]
      summary: Browse recent MQTT messages (last 10k retained)
      parameters:
        - in: query
          name: topic
          schema: { type: string }
          description: 'Topic filter (exact match or prefix with wildcard # e.g. "stat/#")'
        - in: query
          name: device_id
          schema: { type: string, format: uuid }
        - in: query
          name: retained
          schema: { type: boolean }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
        - in: query
          name: before_id
          schema: { type: integer }
          description: Cursor for pagination (return records with id < before_id)
      responses:
        '200':
          description: MQTT message list (newest first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/MqttMessage' }
                  has_more: { type: boolean }
  /mqtt/publish:
    post:
      tags: [MQTT]
      summary: Publish a message to the MQTT broker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MqttPublishRequest'
      responses:
        '200':
          description: Message published
          content:
            application/json:
              schema:
                type: object
                properties:
                  published: { type: boolean }
                  topic:     { type: string }
                  payload:   { type: string }
        '503':
          description: MQTT broker not connected

  # =========================================================================
  # FIRMWARE
  # =========================================================================
...
  /firmware/{device_id}/log-update:
    parameters:
      - in: path
        name: device_id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Firmware]
      summary: Log a firmware update operation (manual or OTA)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirmwareUpdateLog'
      responses:
        '201':
          description: Update record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirmwareUpdateRecord'
  /firmware/{device_id}/history:
    parameters:
      - in: path
        name: device_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Firmware]
      summary: Get firmware update history for a device
      responses:
        '200':
          description: Firmware update history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FirmwareUpdateRecord'

  # =========================================================================
  # DASHBOARD
  # =========================================================================
...